Rename "std" mod to "stdlib"
This avoids confusion with the Rust standard library, and clarifies
(hopefully) that this module contains the implementation of the RCL
standard library.
